GUI: (GSampleNumericalDialog + TemplateNumericalWidget)     - moved class QIodeNumericalObjectWidget as TemplateNumericalWidget in the tab_numerical_values.h header file     - moved class QIodeGSampleTableView as GSampleNumericalDialog in the tab_numerical_values.h header file     - Enabled the possibility for the users to choose the number of digits to display in GSample tables
fix #312
