+++ gui/menu/data/edit_tables.h
@@ -8,8 +8,13 @@
 #include "utils.h"
 #include "settings.h"
 #include "wrapper_classes.h"
-#include "iode_objs/views/gsample_table_view.h"
 #include "custom_widgets/completer.h"
+
+#ifndef GSAMPLE_NUMERICAL_WIDGET_HEADER
+#define _GSAMPLE_
+#include "tabs/tab_numerical_values.h"
+#undef _GSAMPLE_
+#endif
 
 
 /* NOTE FOR THE DEVELOPERS:
@@ -23,7 +28,7 @@ class QIodeMenuDataEditTables: public QIodeSettings, public Ui::QIodeMenuDataEdi
 {
     Q_OBJECT
     
-    QVector<QIodeGSampleTableView*> tableViews;
+    QVector<GSampleNumericalDialog*> tableViews;
 
     WrapperQPlainTextEdit*   wTablesNames;
