+++ gui/menu/print_graph/print_tables.cpp
@@ -36,7 +36,7 @@ QIodeMenuPrintTables::QIodeMenuPrintTables(QWidget* parent) : QIodeMenuPrintAbst
 QIodeMenuPrintTables::~QIodeMenuPrintTables()
 {
     clear_all_reference_kdbs();
-    foreach(QIodeGSampleTableView* view, tableViews) view->close();
+    foreach(GSampleNumericalDialog* view, tableViews) view->close();
     tableViews.clear();
 
     delete wTablesNames;
@@ -82,7 +82,7 @@ void QIodeMenuPrintTables::display()
         for(int i=0; i < nb_names; i++)
         {
             QString tableName(c_table_names[i]); 
-            QIodeGSampleTableView* view = new QIodeGSampleTableView(tableName, gsample, nb_decimals, "", this);
+            GSampleNumericalDialog* view = new GSampleNumericalDialog(tableName, gsample, nb_decimals, "", this);
             tableViews.append(view);
             view->open();
         }
