+++ gui/tabs/tab_database_abstract.h
@@ -26,7 +26,8 @@ protected:
     EnumIodeType iodeType;
     QDir projectDir;
 
-    QGridLayout* layout;
+    int row;
+    QGridLayout* gridLayout;
     QLineEdit* lineEdit_filter;
     QPushButton* pushButton_filter;
     QPushButton* pushButton_print;
@@ -42,9 +43,10 @@ public:
         this->setObjectName(QString::fromUtf8("widget_iode_obj"));
 
         // layout
-        layout = new QGridLayout(this);
-        layout->setObjectName(QString::fromUtf8("layout"));
-        layout->setContentsMargins(0, 0, 0, 0);
+        row = 0;
+        gridLayout = new QGridLayout(this);
+        gridLayout->setObjectName(QString::fromUtf8("layout"));
+        gridLayout->setContentsMargins(0, 0, 0, 0);
 
         // filter 
         lineEdit_filter = new QLineEdit();
@@ -56,20 +58,20 @@ public:
         sizePolicyFilter.setHeightForWidth(lineEdit_filter->sizePolicy().hasHeightForWidth());
         lineEdit_filter->setSizePolicy(sizePolicyFilter);
         lineEdit_filter->setMinimumSize(QSize(200, 0));
-        layout->addWidget(lineEdit_filter, 0, 0, Qt::AlignLeft);
+        gridLayout->addWidget(lineEdit_filter, row, 0, Qt::AlignLeft);
 
         pushButton_filter = new QPushButton("Filter");
         pushButton_filter->setObjectName(QString::fromUtf8("pushButton_filter"));
-        layout->addWidget(pushButton_filter, 0, 1, Qt::AlignLeft);
+        gridLayout->addWidget(pushButton_filter, row, 1, Qt::AlignLeft);
 
         // spacer
         QSpacerItem* horizontalSpacer = new QSpacerItem(828, 20, QSizePolicy::Expanding, QSizePolicy::Minimum);
-        layout->addItem(horizontalSpacer, 0, 2);
+        gridLayout->addItem(horizontalSpacer, row, 2);
 
         // print button
         pushButton_print = new QPushButton("Print");
         pushButton_print->setObjectName(QString::fromUtf8("pushButton_print"));
-        layout->addWidget(pushButton_print, 0, 3, Qt::AlignLeft);
+        gridLayout->addWidget(pushButton_print, row, 3, Qt::AlignLeft);
 
         // add button
         pushButton_add = new QPushButton("Add " + QString::fromStdString(vIodeTypes[iodeType]));
@@ -79,7 +81,9 @@ public:
         sizePolicyAdd.setVerticalStretch(0);
         sizePolicyAdd.setHeightForWidth(pushButton_add->sizePolicy().hasHeightForWidth());
         pushButton_add->setSizePolicy(sizePolicyAdd);
-        layout->addWidget(pushButton_add, 0, 4, Qt::AlignRight);
+        gridLayout->addWidget(pushButton_add, row, 4, Qt::AlignRight);
+
+        row++;
     }
 
     ~AbstractIodeObjectWidget()
@@ -87,7 +91,7 @@ public:
         delete lineEdit_filter;
         delete pushButton_filter;
         delete pushButton_add;
-        delete layout;
+        if(gridLayout) delete gridLayout;
     }
 
     /**
@@ -217,7 +221,8 @@ public:
 
         // insert table to layout
         // -1 -> span over all rows/columns
-        layout->addWidget(tableview, 1, 0, 1, -1);
+        gridLayout->addWidget(tableview, row, 0, 1, -1);
+        row++;
     }
 
     ~QIodeObjectWidget()
